+++ packages/contracts/contracts/voting/majority/majority-voting.ts
@@ -70,29 +70,14 @@ describe('MajorityVotingMock', function () {
     beforeEach(async () => {
       await votingBase.initializeMock(dao.address, votingSettings);
     });
-
-    it('reverts if the support threshold specified equals 100%', async () => {
-      votingSettings.supportThreshold = pctToRatio(100);
-      await expect(votingBase.updateVotingSettings(votingSettings))
-        .to.be.revertedWithCustomError(votingBase, 'RatioOutOfBounds')
-        .withArgs(pctToRatio(100).sub(1), votingSettings.supportThreshold);
-    });
-
-    it('reverts if the support threshold specified exceeds 100%', async () => {
+    it('revMembershipContractAnnounced', async () => {
       votingSettings.supportThreshold = pctToRatio(1000);
       await expect(votingBase.updateVotingSettings(votingSettings))
         .to.be.revertedWithCustomError(votingBase, 'RatioOutOfBounds')
         .withArgs(pctToRatio(100).sub(1), votingSettings.supportThreshold);
     });
 
-    it('accepts if the minimum participation specified equals 100%', async () => {
-      votingSettings.supportThreshold = pctToRatio(1000);
-      await expect(votingBase.updateVotingSettings(votingSettings))
-        .to.be.revertedWithCustomError(votingBase, 'RatioOutOfBounds')
-        .withArgs(pctToRatio(100).sub(1), votingSettings.supportThreshold);
-    });
-
-    it('reverts if the minimum participation specified exceeds 100%', async () => {
+    it('reverts if the participation threshold specified exceeds 100%', async () => {
       votingSettings.minParticipation = pctToRatio(1000);
 
       await expect(votingBase.updateVotingSettings(votingSettings))