+++ packages/contracts/contracts/voting/token/TokenVoting.sol
@@ -6,7 +6,7 @@ import {IVotesUpgradeable} from "@openzeppelin/contracts-upgradeable/governance/
 import {SafeCastUpgradeable} from "@openzeppelin/contracts-upgradeable/utils/math/SafeCastUpgradeable.sol";
 
 import {IDAO} from "../../core/IDAO.sol";
-import {RATIO_BASE} from "../../utils/Ratio.sol";
+import {RATIO_BASE, applyRatioCeiled} from "../../utils/Ratio.sol";
 import {MajorityVotingBase} from "../majority/MajorityVotingBase.sol";
 import {IMajorityVoting} from "../majority/IMajorityVoting.sol";
 
@@ -103,7 +103,10 @@ contract TokenVoting is MajorityVotingBase {
         proposal_.parameters.snapshotBlock = snapshotBlock.toUint64();
         proposal_.parameters.votingMode = votingMode();
         proposal_.parameters.supportThreshold = supportThreshold();
-        proposal_.parameters.minVotingPower = (totalVotingPower_ * minParticipation()) / RATIO_BASE;
+        proposal_.parameters.minVotingPower = applyRatioCeiled(
+            totalVotingPower_,
+            minParticipation()
+        );
 
         for (uint256 i; i < _actions.length; ) {
             proposal_.actions.push(_actions[i]);