+++ packages/contracts/contracts/voting/token/TokenVotingSetup.sol
@@ -33,8 +33,8 @@ contract TokenVotingSetup is PluginSetup {
     /// @notice The address of the `TokenVoting` base contract.
     TokenVoting private immutable tokenVotingBase;
 
-    /// @notice The address zero to be used as oracle address for permissions.
-    address private constant NO_ORACLE = address(0);
+    /// @notice The address zero to be used as condition address for permissions.
+    address private constant NO_CONDITION = address(0);
 
     /// @notice The address of the `GovernanceERC20` base contract.
     address public immutable governanceERC20Base;
@@ -92,11 +92,14 @@ contract TokenVotingSetup is PluginSetup {
     /// @inheritdoc IPluginSetup
     function prepareInstallationDataABI() external pure returns (string memory) {
         return
-            "(tuple(uint8 votingMode, uint64 supportThreshold, uint64 minParticipation, uint64minDuration, uint256 minProposerVotingPower) votingSettings, tuple(address addr, string name, string symbol) tokenSettings, tuple(address[] receivers, uint256[] amounts) mintSettings)";
+            "(tuple(uint8 votingMode, uint32 supportThreshold, uint32 minParticipation, uint64minDuration, uint256 minProposerVotingPower) votingSettings, tuple(address addr, string name, string symbol) tokenSettings, tuple(address[] receivers, uint256[] amounts) mintSettings)";
     }
 
     /// @inheritdoc IPluginSetup
-    function prepareInstallation(address _dao, bytes memory _data)
+    function prepareInstallation(
+        address _dao,
+        bytes memory _data
+    )
         external
         returns (
             address plugin,
@@ -194,7 +197,7 @@ contract TokenVotingSetup is PluginSetup {
             PermissionLib.Operation.Grant,
             plugin,
             _dao,
-            NO_ORACLE,
+            NO_CONDITION,
             tokenVotingBase.UPDATE_VOTING_SETTINGS_PERMISSION_ID()
         );
 
@@ -202,7 +205,7 @@ contract TokenVotingSetup is PluginSetup {
             PermissionLib.Operation.Grant,
             plugin,
             _dao,
-            NO_ORACLE,
+            NO_CONDITION,
             tokenVotingBase.UPGRADE_PLUGIN_PERMISSION_ID()
         );
 
@@ -211,7 +214,7 @@ contract TokenVotingSetup is PluginSetup {
             PermissionLib.Operation.Grant,
             _dao,
             plugin,
-            NO_ORACLE,
+            NO_CONDITION,
             DAO(payable(_dao)).EXECUTE_PERMISSION_ID()
         );
 
@@ -222,7 +225,7 @@ contract TokenVotingSetup is PluginSetup {
                 PermissionLib.Operation.Grant,
                 token,
                 _dao,
-                NO_ORACLE,
+                NO_CONDITION,
                 tokenMintPermission
             );
         }
@@ -265,7 +268,7 @@ contract TokenVotingSetup is PluginSetup {
             PermissionLib.Operation.Revoke,
             _plugin,
             _dao,
-            NO_ORACLE,
+            NO_CONDITION,
             tokenVotingBase.UPDATE_VOTING_SETTINGS_PERMISSION_ID()
         );
 
@@ -273,7 +276,7 @@ contract TokenVotingSetup is PluginSetup {
             PermissionLib.Operation.Revoke,
             _plugin,
             _dao,
-            NO_ORACLE,
+            NO_CONDITION,
             tokenVotingBase.UPGRADE_PLUGIN_PERMISSION_ID()
         );
 
@@ -281,7 +284,7 @@ contract TokenVotingSetup is PluginSetup {
             PermissionLib.Operation.Revoke,
             _dao,
             _plugin,
-            NO_ORACLE,
+            NO_CONDITION,
             DAO(payable(_dao)).EXECUTE_PERMISSION_ID()
         );
 
@@ -295,7 +298,7 @@ contract TokenVotingSetup is PluginSetup {
                 PermissionLib.Operation.Revoke,
                 token,
                 _dao,
-                NO_ORACLE,
+                NO_CONDITION,
                 GovernanceERC20(token).MINT_PERMISSION_ID()
             );
         }