+++ packages/contracts/test/test-utils/voting.ts
@@ -24,8 +24,8 @@ export type VotingSettings = {
   minProposerVotingPower: number;
 };
 
-export const pctToRatio = (x: number) =>
-  ethers.BigNumber.from(10).pow(4).mul(x); // 100% => 10**6, 1% => 10**4
+export const RATIO_BASE = ethers.BigNumber.from(10).pow(6); // 100% => 10**6
+export const pctToRatio = (x: number) => RATIO_BASE.mul(x).div(100);
 
 export const ONE_HOUR = 60 * 60;
 export const ONE_DAY = 24 * ONE_HOUR;
@@ -94,4 +94,9 @@ export async function timestampIn(durationInSec: number): Promise<number> {
 
 export async function setTimeForNextBlock(timestamp: number): Promise<void> {
   await ethers.provider.send('evm_setNextBlockTimestamp', [timestamp]);
+}
+
+export function toBytes32(num: number): string {
+  const hex = num.toString(16);
+  return `0x${'0'.repeat(64 - hex.length)}${hex}`;
 }